Add --silent flag to run_tests.py for suppressing Makefile stdout
Sends only STDOUT to os.devnull; this greatly condenses the logs of
Jenkins and Travis CI builds for readability. Will need to determine if
the non-suppressed STDERR is sufficiently helpful for debugging broken
builds via the logs without needing to modify and recommit the test
flags.

May want to consider expanding this flag's action to configure.py output
and the header/footer output of ./athena run commands.