Encapsulate all physics-dependent floors in new EOS class function
Essentially, this combines the pre-W(U) density floor and post-W(U)
pressure/energy floor from the traditional second-order
ConservedToPrimitive() EOS function.

There are no post-W(U) floors in the isothermal hydro and MHD equations
of state.

The new function could be called to apply the post-W(U) floors in the
Newtonian adiabatic equations of state in ConservedToPrimitive(), but
it would be slower since it recalculates kinetic energy, magnetic
pressure, etc., which are already known in ConservedToPrimitive().

This makes it possible for a single definition of the fourth-order EOS
wrapper function ConservedToPrimitiveCellAverage() to exist and be
shared among all Newtonian --eos options (adiabatic vs. isothermal; mhd
vs. hydro).