Use max_digits10 from <limits> whenever using setprecision()
Referencing this member of the numeric_limits class provides a portable
way to get the entire, meaningful numerical precision of the
floating-point values when writing out diagnostics. Automatically
adjusts precision if Athena++ is configured with -float for
single-precision mode, since std::numeric_limits<Real>::max_digits10
= 17 if Real=double=binary64
= 9 if Real=float=binary32

= # of significant digits required to guarantee round-trip
binary -> decimal -> binary conversion. Therefore:

setprecision(max_digits10 - 1) if std::scientific is the format
specifier (since the function specifies the number of digits only after
the decimal point in this case)

setprecision(max_digits10) if std::defaultfloat is the format
specifier (since the function specifies the number of significant digits
after AND before the decimal point in this case)