Reverted to dynamic OpenMP scheduling.
After a long painful struggle, I found that the original implementation by Jim works better. I assumed the dynamic scheduling might be conflicting with MPI and changed it along with the lock implementation and TaskList initialization, but it did not work. In fact changing the lock and TaskList initialization was sufficient.

Note that the performance improvement by 1D pencil arrays on KNL is no as large as on Skylake. This is likely indicating the fact that KNL has a better Byte/Flop ratio.