Use default implicitly-defined destructor + 4x move/copy methods
(for OutputType). Best practice for future C++ standard compatibility,
since user-declared (not only user-defined!) destructor, even if it is a
trivial destructor that is declared only in order to be virtual, should
be accompanied by the other 4x declarations. This is the Rule of Five.

"The generation of the implicitly-defined copy assignment operator is
deprecated (since C++11) if T has a user-declared destructor or
user-declared copy constructor."

Also add comment describing motiviation for "explicit" keyword on
constructor declarations for all OutputType base and derived classes.