Fix masked bug with shearing box + MHD
This part of AverageEMFBoundary() would access uninitialized members of
the neighbor[56] array, and it contained a hidden assumption that
shearing box simulations always are 3D + no refinement.

This did not cause issues for non-shearing-box simulations by relying on
the guarantee that NeighborBlock::shear=false was default initialization
from the user-provided default ctor. But it should not access these
array members.