Default function parameters should be defined in header fn declaration
Defining them in the function definition for NeighborBlock::SetNeighbor
only worked here because this function is only called in bvals_base.cpp.

If the member function is exposed, as it is here as "public", the only
way all callers can know about the default function parameters is if
the values are present in the initial declaration in the header.

Default parameters should NOT be defined in BOTH the .hpp and .cpp
file. Code duplication, and the requirements that the default argument
for a given parameter can never be specified more than once (even with
the same value!). Only allowed to grow the set of default function
parameters for each subsequent function re-declaration or definition.