Added root-level OMP parallelization
The minimum size of the root grid for parallelization is controlled by the Multigrid::minth_ parameter. This value must be set case by case.

Personal note: currently there is very subtle mismatch in the results between OpenMP and MPI. To some extent it is inevitable because OpenMP and MPI use different operation orderings, but it should be kept in mind.